Move modules under renderers
