+++ src/heart/programs/configurations/halloween_2024.py
@@ -1,7 +1,7 @@
 from heart.display.color import Color
 from heart.environment import GameLoop
-from heart.modules.life.renderer import Life
 from heart.renderers.mandelbrot.scene import MandelbrotMode
+from heart.renderers.modules.life.renderer import Life
 from heart.renderers.pacman import PacmanGhostRenderer
 from heart.renderers.pixels import Border
 from heart.renderers.random_pixel import RandomPixel
