+++ src/heart/programs/configurations/life.py
@@ -1,5 +1,5 @@
 from heart.environment import GameLoop
-from heart.modules.life.renderer import Life
+from heart.renderers.modules.life.renderer import Life
 
 
 def configure(loop: GameLoop) -> None:
