+++ src/heart/renderers/led_wave_boat/provider.py
@@ -8,14 +8,14 @@ import reactivex
 from pygame.time import Clock
 from reactivex import operators as ops
 
-from heart.modules.devices.acceleration.provider import \
-    AllAccelerometersProvider
 from heart.peripheral.core.manager import PeripheralManager
 from heart.peripheral.core.providers import ObservableProvider
 from heart.peripheral.sensor import Acceleration
 from heart.renderers.led_wave_boat.state import (LedWaveBoatFrameInput,
                                                  LedWaveBoatState,
                                                  SprayParticle)
+from heart.renderers.modules.devices.acceleration.provider import \
+    AllAccelerometersProvider
 
 HULL_DEPTH = 1.0
