+++ src/heart/renderers/modules/__init__.py
@@ -1,0 +1,1 @@
+"""Shared renderers bundled as modules with providers and assets."""
