+++ src/heart/renderers/modules/devices/acceleration/__init__.py
@@ -1,5 +1,5 @@
-from heart.modules.devices.acceleration.provider import \
-    AllAccelerometersProvider
 from heart.peripheral.core.providers import container
+from heart.renderers.modules.devices.acceleration.provider import \
+    AllAccelerometersProvider
 
 container[AllAccelerometersProvider] = AllAccelerometersProvider
