+++ src/heart/renderers/modules/devices/switch/__init__.py
@@ -1,4 +1,4 @@
-from heart.modules.devices.switch.provider import MainSwitchProvider
 from heart.peripheral.core.providers import container
+from heart.renderers.modules.devices.switch.provider import MainSwitchProvider
 
 container[MainSwitchProvider] = MainSwitchProvider
